requirements_yolo.txt
# YOLO и Computer Vision
ultralytics
opencv-python
opencv-contrib-python

# Распознавание лиц
face-recognition
dlib-bin

# Работа с данными
numpy>=1.26.0
pandas>=2.0.0
pillow>=10.0.0

# Утилиты
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Для отчетов
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Веб-интерфейс Streamlit
streamlit>=1.28.0
plotly>=5.18.0
openpyxl>=3.1.0

# Веб-интерфейс (опционально)
streamlit>=1.28.0
